fixed solution of this problem(int -> long)

+++ Improving_academic_performance/Improving_academic_performance.java
@@ -10,7 +10,7 @@ public class Improving_academic_performance {
         double c = sc.nextDouble();
 
 
-        int x = (int)Math.ceil((3 * a + b - c) / 3.0);
+        long x = (long)Math.ceil((3 * a + b - c) / 3.0);
         if (a + b + c == 0) {
             x = 1;
         } else if (x < 0) {
